V0.99.7 1.Fix a link bug in admin catalog list with multi subweb 2.Fix a bug in multi subweb cache delete 3.Fix a bug in file cache clear function 4.Fix a bug in plugin install function 5.Fix a cache page bug in article list page when a prefix exists 6.Add subweb selection in list page of news_mark plugin 7.Add subweb selection in list page of news_visit plugin 8.Visit counter will not make count when a spider visit 9.Add delect all link of any topic function in topic plugin 10.Add referer item in visit keyword record of visit analysis plugin 11.Fix a bug in read page when html cache enable 12.Fix a xml format bug in template module 13.Session will not work count when a spider visit 14.Fix a bug in admin log function 15.Fix a bug in getSubSetting function 16.Comment management plugin added 17.Some javascript adjust... 18.Some css adjusts... 19.Some language adjusts... 20.Some other bug fixes and adjusts...

+++ plugin/meeting/class.php
@@ -25,6 +25,9 @@ class plugin_meeting implements plugin {
 			</pre>
 			");
 		} else {
+			WriteFile(dirname(__FILE__)."/config.php", "<?php
+\$catid = {$new_id};
+?>", "wb");
 			deleteCache("admin_cat");
 			includeCache("admin_cat");
 			$admin_cat = toJson($admin_cat, $setting['gen']['charset']);
@@ -36,10 +39,7 @@ parent.setNav();
 mystep;
 			deleteCache("plugin");
 			buildParaList("plugin");
-			WriteFile(dirname(__FILE__)."/config.php", '<?php
-$catid = '.$new_id.';
-?>', "wb");
-			showInfo($setting['language']['plugin_install_done']);
+			echo showInfo($setting['language']['plugin_install_done'], false);
 		}
 	}
 	
@@ -63,7 +63,7 @@ $catid = '.$new_id.';
 		$db->Free();
 		$db->BatchExec($sql_list);
 		include("config.php");
-		$db->query("delete from ".$setting['db']['pre']."admin_cat where pid='".$catid."'");
+		if(isset($catid) && $catid!=0)	$db->query("delete from ".$setting['db']['pre']."admin_cat where pid='".$catid."'");
 		$db->query("truncate table ".$setting['db']['pre']."meeting");
 		$db->query("drop table ".$setting['db']['pre']."meeting");
 		$db->query("delete from ".$setting['db']['pre']."admin_cat where name='ª·“È'");
@@ -91,7 +91,7 @@ mystep;
 			WriteFile(dirname(__FILE__)."/config.php", '<?php
 $catid = 0;
 ?>', "wb");
-			showInfo($setting['language']['plugin_uninstall_done']);
+			echo showInfo($setting['language']['plugin_uninstall_done'], false);
 		}
 	}
